# feat: comprehensive performance optimizations - 97% startup improvement