--- a/confluence_gateway/cli/index_commands.py
+++ b/confluence_gateway/cli/index_commands.py
@@ -2,14 +2,11 @@ import logging
 
 import typer
 
-from confluence_gateway.api.schemas.responses import IndexingStatusResponse
 from confluence_gateway.cli.common import (
     handle_cli_errors,
     print_indexing_status,
     print_status,
 )
-from confluence_gateway.cli.dependencies import _get_indexing_service
-from confluence_gateway.services.indexing import IndexingService
 
 logger = logging.getLogger(__name__)
 
@@ -36,6 +33,9 @@ def trigger_indexing(
         help="Index all accessible spaces (ignores configuration filters).",
     ),
 ) -> None:
+    from confluence_gateway.cli.dependencies import _get_indexing_service
+    from confluence_gateway.services.indexing import IndexingService
+
     indexing_service: IndexingService | None = _get_indexing_service()
 
     if indexing_service is None:
@@ -88,6 +88,10 @@ def trigger_indexing(
 @app.command("status", help="Get the current status of the indexing service.")
 @handle_cli_errors
 def get_status() -> None:
+    from confluence_gateway.api.schemas.responses import IndexingStatusResponse
+    from confluence_gateway.cli.dependencies import _get_indexing_service
+    from confluence_gateway.services.indexing import IndexingService
+
     indexing_service: IndexingService | None = _get_indexing_service()
 
     if indexing_service is None: